fix(db) delete expired rows without transaction with postgres
The reason is that transaction may acquire unneccessary locks, and is
also non-resilient to errors, e.g. if one table is missing `ttl` column
no cleanups will be done to other tables either.

This changes the code so that cleanups are ran separately on each table.

It might have something to do with #4601.